refactor: extract localized UI label extensions
